pyproject.toml
[tool.poetry]
name = "dragonjudge"
version = "0.1.0"
description = "A股情绪龙头助手 - 智能分析系统"
authors = ["zhdtty <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.0"
numpy = "^1.26.0"
aiohttp = "^3.9.0"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
schedule = "^1.2.0"
pydantic = "^2.6.0"
python-dotenv = "^1.0.0"
loguru = "^0.7.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.26.0"

[tool.poetry.dev-dependencies]
pytest-cov = "^4.1.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short --cov=src --cov-report=term-missing"
